well

+++ app/main.py
@@ -1,5 +1,5 @@
 class Cargo:
-    def __init__(self, weight: int) -> None:
+    def __init__(self, weight: float) -> None:
         self.weight = weight
 
 class BaseRobot:
